requirements.txt
Django==4.0.3
Pillow
asgiref==3.7.2
django==3.2.10 mysqlclient
dnspython==2.4.2
sqlparse==0.2.4
tzdata==2023.3

